+++ backend/src/main/java/first/webide/dto/request/Member/DeleteMemberRequest.java
@@ -1,4 +1,4 @@
-package first.webide.dto.request;
+package first.webide.dto.request.Member;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
